requirements.txt
# Core FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Database (Async PostgreSQL)
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Caching (Redis with async support)
redis>=5.0.1

# Task Scheduling
apscheduler>=3.10.4

# AI & Text-to-Speech
gtts>=2.3.2
google-generativeai>=0.3.0


# OpenAI Python SDK
openai>=0.28.0

# ElevenLabs TTS SDK
elevenlabs>=0.2.0

